restaubot/requirements.txt
psycopg
psycopg2-binary
django
djangorestframework
Pillow
drf-yasg
django-cors-headers
clerk-django